Move models to models package
